pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = "Finange Bot"
authors = ["The Hack's"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.1"
python-telegram-bot = "^13.14"
pytest = "^7.1.3"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
pycodestyle = "^2.9.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
